move generic task creation methods to controller:: as they are not specific to Talos
