+++ include/inria_wbc/controllers/talos_pos_tracking.hpp
@@ -1,17 +1,6 @@
 #ifndef IWBC_POS_TRACKING_HPP
 #define IWBC_POS_TRACKING_HPP
 
-#include <tsid/contacts/contact-6d.hpp>
-#include <tsid/contacts/contact-point.hpp>
-#include <tsid/math/utils.hpp>
-#include <tsid/tasks/task-actuation-bounds.hpp>
-#include <tsid/tasks/task-com-equality.hpp>
-#include <tsid/tasks/task-joint-bounds.hpp>
-#include <tsid/tasks/task-joint-posVelAcc-bounds.hpp>
-#include <tsid/tasks/task-joint-posture.hpp>
-#include <tsid/tasks/task-se3-equality.hpp>
-#include <tsid/trajectories/trajectory-base.hpp>
-
 #include <inria_wbc/controllers/controller.hpp>
 
 namespace inria_wbc {
@@ -60,15 +49,7 @@ namespace inria_wbc {
             virtual void parse_configuration_yaml(const std::string& sot_config_path);
             virtual void set_stack_configuration();
             virtual void set_default_opt_params(std::map<std::string, double>& p);
-
             std::shared_ptr<tsid::contacts::Contact6d> make_contact_task(const std::string& name, const std::string frame_name, double kp) const;
-            std::shared_ptr<tsid::tasks::TaskComEquality> make_com_task(const std::string& name, double kp) const;
-            std::shared_ptr<tsid::tasks::TaskJointPosture> make_posture_task(const std::string& name, double kp) const;
-            std::shared_ptr<tsid::tasks::TaskSE3Equality> make_torso_task(const std::string& name, const std::string& frame_name, double kp) const;
-            std::shared_ptr<tsid::tasks::TaskSE3Equality> make_floatingb_task(const std::string& name, const std::string& joint_name, double kp) const;
-            std::shared_ptr<tsid::tasks::TaskSE3Equality> make_hand_task(const std::string& name, const std::string& joint_name, double kp) const;
-            std::shared_ptr<tsid::tasks::TaskSE3Equality> make_foot_task(const std::string& name, const std::string& joint_name, double kp) const;
-            std::shared_ptr<tsid::tasks::TaskJointPosVelAccBounds> make_bound_task(const std::string& name) const;
 
             std::map<std::string, double> opt_params_; // the parameters that we can tune with an optimizer (e.g., task weights)
 
@@ -87,23 +68,6 @@ namespace inria_wbc {
             // bounds
             std::shared_ptr<tsid::tasks::TaskJointPosVelAccBounds> bounds_task_;
         };
-
-        inline tsid::trajectories::TrajectorySample to_sample(const Eigen::VectorXd& ref)
-        {
-            tsid::trajectories::TrajectorySample sample;
-            sample.pos = ref;
-            sample.vel.setZero(ref.size());
-            sample.acc.setZero(ref.size());
-            return sample;
-        }
-
-        inline tsid::trajectories::TrajectorySample to_sample(const pinocchio::SE3& ref)
-        {
-            tsid::trajectories::TrajectorySample sample;
-            sample.resize(12, 6);
-            tsid::math::SE3ToVector(ref, sample.pos);
-            return sample;
-        }
 
     } // namespace controllers
 } // namespace inria_wbc
